Fixed - the implementation of the complete form to see the results

+++ include/coinsExchange.h
@@ -28,7 +28,7 @@ class coinsExchange {
     void setAmount(double amount);
   private:
     std::vector<double> coins_; /// Este es el conjunto de candidatos de monedas viables.
-    float amount_; /// Este es el valor objetivo.
-    float addition_; /// Este es el valor de la suma de las monedas seleccionadas.
+    double amount_; /// Este es el valor objetivo.
+    double addition_; /// Este es el valor de la suma de las monedas seleccionadas.
     std::vector<double> selected_; /// Este es el conjunto de soluciones.
 };
